# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aerobay Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "List of categories",
                "operationId": "fda83319ea87b70fdda27e40a948692b",
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Scaner"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Some description"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create new category",
                "operationId": "5befb384353e20ca29b5464c59701011",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Scaner"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Some description"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Scaner"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Some description"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/categories/{category}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Individual category",
                "operationId": "d119cf95b886ef315d9883e0161bc2ee",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID of category",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Scaner"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Some description"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update category data",
                "operationId": "b32862b5571abe03cf0b1fced59832cf",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID of category",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Scaner"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Some description"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Scaner"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Some description"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete category",
                "operationId": "aa760339c7ddb217329ab434247bedec",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "ID of category",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
...
    }
}